pyproject.toml
[project]
name = "projet-traitement-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.5.0",
    "mpmath>=1.3.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pandas>=1.3.0",
    "pillow>=8.0.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.0.0",
    "seaborn>=0.11.0",
    "timm>=1.0.22",
    "tqdm>=4.62.0",
]

